requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.35.0
